[Length Deprecation] Remove unused LengthSize and LengthBox types https://bugs.webkit.org/show_bug.cgi?id=299239
Reviewed by Darin Adler.

Removes the `LengthSize` and `LengthBox` types.

LengthBox.h had a few unrelated types in it that should have instead
been in BoxExtents.h, so they have been moved there.

* Source/WebCore/Headers.cmake:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/layout/formattingContexts/table/TableGrid.h:
* Source/WebCore/page/PrintContext.cpp:
* Source/WebCore/page/PrintContext.h:
* Source/WebCore/platform/BoxExtents.cpp: Renamed from Source/WebCore/platform/LengthBox.cpp.
* Source/WebCore/platform/BoxExtents.h:
* Source/WebCore/platform/LengthBox.h: Removed.
* Source/WebCore/platform/LengthFunctions.cpp:
* Source/WebCore/platform/LengthFunctions.h:
* Source/WebCore/platform/LengthSize.cpp: Removed.
* Source/WebCore/platform/LengthSize.h: Removed.
* Source/WebCore/platform/adwaita/ThemeAdwaita.cpp:
* Source/WebCore/platform/graphics/FloatRect.h:
* Source/WebCore/platform/graphics/FontCascade.h:
* Source/WebCore/platform/graphics/LayoutRect.h:
* Source/WebCore/platform/graphics/avfoundation/objc/MediaSampleAVFObjC.h:
* Source/WebCore/platform/graphics/ca/TileController.h:
* Source/WebCore/platform/graphics/controls/ControlStyle.h:
* Source/WebCore/platform/graphics/filters/FilterFunction.h:
* Source/WebCore/platform/graphics/mac/controls/ControlMac.h:
* Source/WebCore/rendering/CSSFilterRenderer.h:
* Source/WebCore/rendering/NinePieceImagePainter.cpp:
* Source/WebCore/rendering/style/BorderData.h:
* Source/WebCore/rendering/style/RenderStyle.h:
* Source/WebCore/rendering/style/StyleRareNonInheritedData.h:
* Source/WebCore/rendering/style/StyleScrollSnapPoints.h:
* Source/WebCore/style/StyleBuilderCustom.h:
* Source/WebCore/style/StyleInterpolationFunctions.h:
* Source/WebCore/style/values/backgrounds/StyleBackgroundLayer.h:
* Source/WebCore/style/values/backgrounds/StyleFillLayers.h:
* Source/WebCore/style/values/masking/StyleMaskLayer.h:
* Source/WebKit/Shared/PrintInfo.h:
* Source/WebKit/Shared/VisibleContentRectUpdateInfo.cpp:
* Source/WebKit/Shared/VisibleContentRectUpdateInfo.h:
* Source/WebKit/Shared/WebCoreArgumentCoders.serialization.in:
* Source/WebKit/UIProcess/API/Cocoa/WKWebViewInternal.h:
* Source/WebKit/UIProcess/WebPageProxy.cpp:
* Source/WebKit/WebProcess/FullScreen/WebFullScreenManager.h:

Canonical link: https://commits.webkit.org/300295@main
