+++ Source/WebCore/platform/BoxExtents.h
@@ -27,11 +27,37 @@
 
 #include <WebCore/LayoutUnit.h>
 #include <WebCore/RectEdges.h>
+#include <optional>
 
 namespace WebCore {
 
 using FloatBoxExtent = RectEdges<float>;
 using IntBoxExtent = RectEdges<int>;
+using IntOutsets = IntBoxExtent;
 using LayoutBoxExtent = RectEdges<LayoutUnit>;
+using LayoutOptionalOutsets = RectEdges<std::optional<LayoutUnit>>;
+
+inline LayoutBoxExtent toLayoutBoxExtent(const IntBoxExtent& extent)
+{
+    return {
+        extent.top(),
+        extent.right(),
+        extent.bottom(),
+        extent.left()
+    };
+}
+
+inline FloatBoxExtent toFloatBoxExtent(const IntBoxExtent& extent)
+{
+    return {
+        static_cast<float>(extent.top()),
+        static_cast<float>(extent.right()),
+        static_cast<float>(extent.bottom()),
+        static_cast<float>(extent.left()),
+    };
+}
+
+WTF::TextStream& operator<<(WTF::TextStream&, const IntBoxExtent&);
+WEBCORE_EXPORT WTF::TextStream& operator<<(WTF::TextStream&, const FloatBoxExtent&);
 
 } // namespace WebCore
