+++ Source/WebCore/style/values/backgrounds/StyleBackgroundLayer.h
@@ -26,7 +26,6 @@
 #pragma once
 
 #include "GraphicsTypes.h"
-#include "LengthSize.h"
 #include "RenderStyleConstants.h"
 #include "StyleBackgroundSize.h"
 #include "StyleFillLayers.h"
