+++ Source/WebCore/style/values/masking/StyleMaskLayer.h
@@ -26,7 +26,6 @@
 #pragma once
 
 #include "GraphicsTypes.h"
-#include "LengthSize.h"
 #include "RenderStyleConstants.h"
 #include "StyleBackgroundSize.h"
 #include "StyleFillLayers.h"
