+++ Source/WebKit/Shared/VisibleContentRectUpdateInfo.cpp
@@ -29,7 +29,6 @@
 
 #include "VisibleContentRectUpdateInfo.h"
 
-#include <WebCore/LengthBox.h>
 #include <wtf/text/TextStream.h>
 
 namespace WebKit {
